packages/jira-ticket-sync/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cptools-jira-ticket-sync"
version = "0.1.0"
description = "Manifest-driven Jira ticket sync CLI"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "PyYAML>=6.0",
    "cptools-common @ git+ssh://git@github.com/cpoopc/share-libs.git@main#subdirectory=tools/python/libs/common",
    "cptools-jira @ git+ssh://git@github.com/cpoopc/share-libs.git@main#subdirectory=tools/python/libs/jira",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
jira-ticket-sync = "cptools_jira_ticket_sync.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cptools_jira_ticket_sync*"]

[tool.setuptools.package-data]
cptools_jira_ticket_sync = [
    "assets/workspace/.env.example",
    "assets/workspace/project-config.yaml",
    "assets/workspace/manifests/README.md",
    "assets/workspace/profiles/*.yaml",
    "assets/workspace/templates/imported/*.yaml",
    "assets/workspace/state/*.json",
    "assets/workspace/state/field-classification/*.json",
]

[tool.uv.sources]
cptools-common = { path = "../../tools/python/libs/common", editable = true }
cptools-jira = { path = "../../tools/python/libs/jira", editable = true }
